retain snapshot failure reason
publication manager is now able to terminate shapes while requests are
waiting for the snapshot to start, meaning that if the reason for
failure arises from issues with e.g. permissions in the publication
manager then we fail to propagate those to the waiting clients.

this restores the snapshot error propagation in the case of publication
perms errors.