feat(sync-service): Re-enable pluggable snapshot function
Phoenix.Sync relies on patching the snapshot function in its sandbox
repo feature (where it uses an Ecto.Repo instance to get a pool
connection and returns fake xids). When we removed dependency injection
we broke this, so this PR just adds back in the ability to inject a
custom stream_snapshot_from_db/5 function.

I also removed the requirement to pass seed_config to the stack config
instance by adding in defaults for most of the required values and
removed the unused `shape_changes_registry` value from the stack config.
We have defaults for these values so we should use them. Generally if we
add config we should have a default -- it makes Phoenix.Sync less likely
to break with every addition.

In DynamicConsumerSupervisor I noticed a potential problem where the
DynamicSupervisor spec passed to PartitionSupervisor could have a `name`
which would cause problems.