feat: Consumers idempotently handle transactions (#3452)
Closes https://github.com/electric-sql/electric/issues/3447

This is part of a larger piece of work described in
https://github.com/electric-sql/electric/issues/3442

We were already filtering transactions at the SLC level, but for the
sake of loosening the requirements of having a central "latest LSN
processed" node for filtering we move the filter at the consumer level.

For the sake of maintaining what we already have in place, I'm simply
comparing the latest log offset of the transaction compared to the
latest log offset tracked by the consumer - the flush acknowledgements
are separate anyway, if a flush ack is not sent for a transaction then
it will be sent again anyway.

I've also fixed the `latest_offset` state variable kept in the consumer
which was not being updated.