feat(sync-service): Send individual operations to ShapeLogCollector (#3429)
Part of #3414, this PR sends individual operations from the
ReplicationClient to ShapeLogCollector. This should reduce the working
memory used by the ReplicationClient, but the ultimate aim is to send
individual operations to storage and so not buffer at any point (unless
subqueries are involved). This PR just moves the transaction
consolidation one step, from ReplicationClient to ShapeLogCollector to
keep the PR small. Other PRs will move the consolidation further, from
ShapeLogCollector to Consumer, then finally from to Consumer to storage
(except with subqueries are involved).

Currently the replication stream take two forms:
1. `Replication.Messages` such as `Messages.Begin`, `Messages.Commit`,
`Messages.Insert` etc.
2. `Changes.Transaction` (or `Changes.Relation`)

And `Collector` converts one to the other.

This PR introduces an intermediate form, Operations, as Messages are
coupled to the postgres replication protocol and are not optimised for
use with Shapes (for example they do not include the relation name) and
Transactions are whole transactions. So with this PR we have:
1. `Replication.Messages` such as `Messages.Begin`, `Messages.Commit`,
`Messages.Insert` etc.
2. Operation (spec: `Changes.operation()`) such as `Changes.Begin`.
`Changes.Commit`, `Changes.NewRecord` etc.
4. `Changes.Transaction` (or `Changes.Relation`) which I'm calling
database actions (spec: `Changes.action()`)

`Collector` has been replaced by `MessageConverter` that converts 1 to 2
and `TransactionBuilder` that coverts 2 to 3.

Benchamarking this PR validated @icehaunter 's hypothesis that extra
messages would slow down small transactions, so this PR also includes a
`OperationBatcher` that batches operations, which the benchmarks confirm
eliminate the slow down.