feat: LSN tracker persistence through replication slot confirmed flush position (#3462)
Closes https://github.com/electric-sql/electric/issues/3448

This is part of a larger piece of work described in
https://github.com/electric-sql/electric/issues/3442

The initial idea was to have a separate, periodic persister of the last
processed LSN, but as I was working through the code I realised that we
might not even need that.

Since now consumers are idempotent in the transactions they handle (see
https://github.com/electric-sql/electric/pull/3452), even the our global
last processed LSN is slightly behind any given shape that is ok.

Instead of persisting it, we actually recover it from the replication
slot's confirmed flush LSN.
- If the slot is new, we reset the last processed LSN to the new slot's
LSN since all shapes will be invalidated
- If the slot is an existing one, we initialise it to the confirmed
flush LSN
- If the reconnection to the slot happens at runtime (e.g. because of
network partition), we preserve the current last processed LSN
- If the reconnection happens at startup, we initialise the LSN from the
slot. This value might be _slightly_ behind if shapes processed
transactions that they did not manage to acknowledge back, but we can
tolerate this.

This way, shape recovery does not have to read all shapes' last
processed LSN during startup.

If we need stronger guarantees of the global processed lsn being the
max, we can also attempt to set it everytime we lazily recover an
existing shape (if the shape's max lsn is larger than the current max
processed lsn). Happy to include this in this PR or a next one.