expose ShapeStream.fetchSnapshot as public api (#3463)
This exposes the `ShapeStream.fetchSnapshot` as a public api that can be
used to fetch a snapshot without it being injected into the emitted
stream of change messages. This is useful for cases where the user wants
to handle the application of these snapshot in a custom way, as we need
to in Tanstack DB for the `progressive` mode.

While working on this I also took the opportunity to tidy up the parsing
of the messages, in the old implementation we would JSON.stringify the
messages, only to the parse them again. I refactored the parser to be
able to work on already parsed (JSON -> basic JS) messages.

First commit is the core change to expose `fetchSnapshot`, second commit
is the refactor of message parsing.