feat: Separate `shape def -> handle` table from metadata (#3466)
Closes https://github.com/electric-sql/electric/issues/3450
Fixes https://github.com/electric-sql/electric/issues/3456 as a bonus

This is part of a larger piece of work described in
https://github.com/electric-sql/electric/issues/3442

I have separated the `shape def -> handle` lookup into its own table -
this alone is not necessarily an improvement but it is the precursor to
lazily loading the shape metadata table as requests or data for shapes
comes in.

In order to split up the work, despite separating the tables, I'm still
backing up both of them to maintain the same functionality as before. In
the next steps I will only be backing up the hash lookup.

As part of this work I have also ensured the last used table is
repopulated upon restoring from the backup with new dates - I don't
think it's worth backing that table up - and I have replaced the `count`
API to use the hash lookup table instead.