chore: Clean up `ShapeStatus` implementation (#3471)
This work is in preparation of
https://github.com/electric-sql/electric/issues/3451

- Removes `set_snapshot_xmin` and `snapshot_xmin` API as it is basically
unused
- Convert `snapshot_started` metadata into positional entry in metadata
tuple (and rename setter API to `mark_snapshot_as_started` for
consistency)
- ETS table names only accept options with a stack id or a stack id,
rather than an arbitrary table name, eliminating the possibility for
issues like the one encountered for the LRU cache @robacourt
- Remove table names from public API and behaviour - this is entirely
internal now
- Make listing shapes for given relations a part of the behaviour
- Fix up tests
  - Remove unused tests
  - Use test stack IDs where possible rather than arbitrary table names 
  - Use stub inspector rather than arbitrary module functions
  
  
I considered (and in fact have a commit stashed for it) splitting the
relation lookup table from the metadata table, into a table of type
`:bag`, but then realised that it would also have to be an
`:ordered_set` since for a given relation we might have an arbitrary
number of shapes attached to it, and we need ordering on the key for
efficient per-shape removals.

The only argument I still potentially have for splitting it out is that
the relation lookups have a very different access pattern - we add
entries when a shape is created, we look them up only for the sake of
shape deletions, we remove them when a shape is removed. The read
pattern is exclusively range reads on a given relation oid partial key
head. This _might_ create contention on an otherwise critical table.

I would like to hear opinions on splitting it out @alco @robacourt 

edit: I've opened the PR for splitting out the relation lookup table as
I've convinced myself https://github.com/electric-sql/electric/pull/3472