feat: Split out relation to shape lookup ETS table in `ShapeStatus` (#3472)
See discussion in https://github.com/electric-sql/electric/pull/3471

This is also in preparation of
https://github.com/electric-sql/electric/issues/3451

Basically we only use the relation lookup for deleting shapes and we do
large range scans over the table which might lock it up or create
contention, even if the table is partitioned via `:write_concurrency`.
Currently it is colocated with the metadata table which is on the
critical read and write path.

By splitting it into its own table, we ensure that low priority scans
for deleting shapes never end up blocking reads or writes to the
metadata table.

As a bonus, until I actually implement lazy recovery of shape metadata,
by moving this into a separate table we also no longer persist these as
our backup as they can be recovered from the shape itself