fix: Fix connection lock time out tests (#3478)
Followup to https://github.com/electric-sql/electric/pull/3476

This is the _actual_ fix. I've reproduced the flakes locally purely
using a stress simulator (should've followed my own advice...) and after
extensive debugging and lots of gaslighting by bots, the issue is that
our test db pool is 1) linked to the test process and 2) set to never
restart connections and instead die, failing the test.

Because this test explicitly kills off a connection with a timeout,
depending on the timing of the pool dying it fails. Unlinking the pool
itself for the test does the trick.