fix: Avoid edge case with 409 causing more SSE requests (#3477)
Fix test flake (see [this
run](https://github.com/electric-sql/electric/actions/runs/19533935381/job/55923347981)
for example) caused by 409s for the shape resetting the fallback from
SSE to long polling.

Not sure if 409s _should_ reset this fallback but it does, and the test
should account for it as we occasionally see these 409s from the tests
overlapping each other.