fix: Shape comparable should include log mode to avoid collisions (#3486)
Fixes https://github.com/electric-sql/electric/issues/3484

What I was worried about happened - the comparable was not updated as we
updated the shape. I think we should have a better more solid way to
define the comparable.

Also updating this will kind of invalidate existing stored shapes if
using shape status backup, but not quite sure what the best way to do
that is.

I've invalidated it by bumping the shape status backup version so this
works now.

---------

Co-authored-by: Valter Balegas <balegas@electric-sql.com>
Co-authored-by: Garry Hill <garry@magnetised.net>