Replace direct telemetry deps with the electric_telemetry library (#3412)
Extract telemetry code from Electric into a separate package. On top of
that, the new package accepts additional user-provided periodic
measurements and metric definitions for otel/prometheus exporters; both
exporters now use the same list of metrics.

One thing that's been lost in this move is the `stack_id` sentry tag
inside the CallHomeRepoter process. Sentry is a dependency of Electric
and there's no straightforward way of propagating it into
CallHomeReporter which is a module in the electric_telemetry dependency
now. I don't feel like it's a big loss because we set stack_id in the
logger metadata for that process, as well as mention it in the process
name and label.

With this merged, we can delete the electric-sql/electric-telemetry
repo.