chore(sync-service): Reduce PublicationManager memory usage (#3496)
Track state using just the table `oid`, not an `{oid, {schema, table}}`
tuple and maintain a separate map of `oid => {schema, table}`.

Also build initial state in an ephemeral task so that we immediately
garbage collect the results of loading the shape status table into RAM.

On a system with 150,000 defined shapes, this reduces the initial memory
usage of PublicationManager from a steady state ~70MB (~270MB before GC)
to a consistent ~21MB.

Also found and fixed a bug when cleaning relations -- we missed an
argument change in the call to ShapeCleaner.cleanup_shapes_async/2 --
and fixed the behaviour of waiting for publication updates when failing
(the start time was re-calculated every attempt, so the test never timed
out).

Fixes #3492