chore(sync-service): Cleanup API struct and GC http handler processes
API instance was carrying a lot of fields it didn't need anymore. This
doesn't affect memory usage unfortunately. Setting fullsweep_after for
the Thousand Island acceptor processes should though -- since they're
re-used they accumulate a lot of binaries. The lower fullsweep_after is
the higher the cpu cost so I've set 10 minor gc sweeps as the default. 0
really cleans up the memory of the processes but is not reasonable for
prod.