chore(sync-service): Hibernate ShapeStatusOwner process (#3505)
If the recovery process is unable to restore the status ets tables from
backup files and has to iterate the filesystem then ShapeStatusOwner is
left with a lot of accumulated memory (~300MB for a system with 200K
shapes).

Release this by hibernating immediately (this process does nothing but
own ets tables anyway).