feat(sync-service): Validate shape handles using quick shape hash validation
This is part 1 of #3511.

Rather than loading a shape handle from the ShapeStatus LUTs using the
Shape definition and then checking that it matches the one in the
request params, we instead do a simple check that the request's handle
and shape definition match using a hash equality check.

We only do a shape def -> handle lookup if the request doesn't have a
handle parameter or the hash check fails

Fixes #3512