chore(sync-service): Refactor ShapeStatus to encapsulate shape data
This is essentially what we have now but with a clear api between fast
metadata lookups and shape -> handle and handle -> shape lookups.

This moves the %Shape{} instances out of the metadata table and gives a
hint of the memory savings to come, this is for ~150K (very simple)
shapes:

```
shapedb:shape_lookup:single_stack    | 149432       | 242.7575 MiB
shapedb:handle_lookup:single_stack   | 149432       | 83.1469 MiB
shape_meta_table:single_stack        | 149432       | 29.5633 MiB
shape_relation_lookup_table:single_s | 149432       | 19.3010 MiB
shape_last_used_table:single_stack   | 149432       | 15.8808 MiB
```

You can see that the shapedb tables are responsible for most of the ets
usage and the metadata now only consumes ~30MiB, so a ~90% reduction,
which means we shouldn't have to go any further and move more things out
of ram.

The current api is not going to work for a sqlite backed version, but
it's close I think.

Fixes #3513