Set the default value for otel_opts in  ElectricTelemetry.Opts' schema
Without the default value, the whole key is omitted from the validated
options. Setting it ensures that the key is included along with all the
default values defined for its subkeys.

We need this to have the default value for export_period override any
default defined in otel_metric_exporter.