feat: Reconcile `ShapeStatus` ETS backup after integrity check (#3527)
Closes https://github.com/electric-sql/electric/issues/3475

Part of a larger piece of work described in
https://github.com/electric-sql/electric/issues/3442

Until now, we were doing a check of what shape handles we have stored
with a cheap `ls` call on the shape directory, and if the handles did
not match the ones found in the ETS backup, we would throw away the
backup.

This is a bit _too_ aggressive, since the ETS backup is actually the
best way to recover quickly, and is bound to fail often if we
periodically back up the tables to hedge against unexpected crashes.

This PR allows for backup reconciliation. Rather than _either_ loading a
full backup that is perfect _or_ loading all shapes from the files, it
allows loading any shape handles not found in the backup from storage.

Basically, we load the ETS backup, we do an `ls` on the shapes dir, any
handles not in the backup but in the storage are loaded from storage,
and any handles in the backup but not in the storage are discarded.