chore(sync-service): Cleanup API struct and GC http handler processes (#3499)
API instance was carrying a lot of fields it didn't need anymore. This
doesn't affect memory usage unfortunately. Setting `fullsweep_after` for
the Thousand Island acceptor processes should though -- since they're
re-used they accumulate a lot of binaries.

The lower `fullsweep_after` is the higher the cpu cost so I've set it to
`10` as the default. `0` really cleans up the memory of the processes
but is not reasonable for prod.

https://www.erlang.org/doc/apps/erts/erlang.html#spawn_opt/4

Update: after the discussion below I've changed this to use
`max_requests` to shutdown http connection handler processes after some
number of requests. I don't think this will have a huge effect on RAM
usage overall but it will deal with the kind of problems we saw after
the CF outage, where request processes accumulated messages