feat: Use ETS table for tracking shapes in RelationTracker (#3546)
Result from the same investigation that lead to
https://github.com/electric-sql/electric/pull/3545

The tracked shapes map is of unbounded size (as large as there are
shapes) - so moved it into an ETS set table instead.

~~I've also removed the tracking of the generated column flag since we
never retrieved it so it was not necessary.~~
EDIT: the generated flag was there for a reason (invalidation in case of
publication change) - reinstated it and updated the code to use it from
the ETS table

This is a simple case of what we were talking about doing on the shape
log collector filters.

We could also use ETS for the rest of the maps for oid -> relation and
relation filters in general, but those are bounded in size by the number
of relations being tracked, which frankly should never reach a number
that is ever high enough to require this sort of move and for it to be a
GC concern.