Report the correct reatined WAL size confirmed_flush_lsn_lag metrics (#3543)
Fix [#3544](https://github.com/electric-sql/electric/issues/3544).

The metric that is currently reported as
`electric.postgres.replication.wal_size` is incorrectly named. Its name
was changed from `electric.postgres.replication.lag` in
https://github.com/electric-sql/electric/pull/2056.

Comparing `pg_current_wal_lsn()` to the replication slot's
`confirmed_flush_lsn` shows how caught up the replcation slot consumer
is to the latest database state. So it does measure the lag in terms of
the total byte size of changes that had been sent to Electric over the
replication connection and how much of those bytes Electric confirmed to
have successfully flushed to disk. This values tends to 0 (or -1 but we
can it at 0) as time progresses and no new changes appear in PG.

However, the other metric that is super critical to Electric's uptime is
the difference between `pg_current_wal_lsn()` and the replication slot's
`restart_lsn`. This shows how much WAL Postgres is forced to keep for
the slot and this is the value that, when it exceeds the
`max_slot_wal_keep_size` config, causes the replication slot to get
invalidated. We have neglected collecting this metric until now.