feat: Batch SLC shape registrations and deregistrations in separate process (#3549)
Follows from discussions on GC and SLC pressure.

The SLC is a very busy process, as it churns through all operations in
the replication stream which can be extremely busy.

At the same time, there is a constant stream of shape registrations and
deregistrations which are currently handled by the SLC as well, which
get interleaved in the log processing. This might be causing lots of
index building and rebuilding in betweeen processing transactions at an
unnecessarily hight rate.

This PR introduces a Registrator process that basically batches shape
registrations and deregistrations, and sends them batched to the SLC
Processor for handling. While it is waiting for the SLC to reply that it
has updated the index, it collects/batches shape registrations and
deregistrations to send as soon as the previous batch is done.

It _could_ be doing a debounce, but I think having it serially update
the index is ensuring that the SLC basically only ever sees 1 message
for updating the index in its mailbox at a time. This basically
prioritizes log consumption over new shape creation and shape removal.

This is also the preparatory work if we want to move in a sort of
out-of-process index updating with deltas being prepared and sent to be
applied much more quickly.