Integration test for Otel metric export (#3553)
Follow-up to https://github.com/electric-sql/electric/pull/3543.

It was a good idea to add a test harness for metric export because in
the process I discovered an issue with the WAL size query:

12:18:52.404 pid=<0.464.0> [warning] Failed to query retained WAL size
Error: ** (Postgrex.Error) ERROR 22003 (numeric_value_out_of_range)
bigint out of range

I have updated the query and the metric name for `pg_wal_offset` since
it no longer matches the value returned by `pg_current_wal_lsn()`. See
explanation in code.

Right now it doesn't matter that the offset stay within the signed
64-bit integer range because our otel_metric_exporter converts values
for the LastValue metric to double. But once
https://github.com/electric-sql/elixir-otel-metric-exporter/pull/28 is
released and Electric upgrades its dep to that version, it will be able
to export integer values as integers.