fix: Recover normalized latest offset for shapes with only a snapshot (#3562)
Fixes https://github.com/electric-sql/electric/issues/3560
Mitigates https://github.com/electric-sql/electric/issues/3551
significantly, but all causes are not addressed by this so will close it
with a separate PR

The issue that has come up again is
https://github.com/electric-sql/electric/issues/2596 which was fixed by
https://github.com/electric-sql/electric/commit/bb4a439e1592b1a24ea207087d83466ffe92b0e8
by normalizing the latest offset recovered.

This is a consequence of the "last before real offsets" value being a
bit of a tricky one to handle and potentially worth mapping out its
semantics a bit more clearly @icehaunter.

Someone coming in with it technically wants to see the log rather than
the snapshot, but we don't persist that the snapshot has "finished"
anywhere, so when we recover the latest offset we have is a strictly
smaller one than `0_inf`.

After introducing lazy shape initialization, which meant that reading a
shape does not require a consumer to have been started which would
trigger the previously implemented fix, the recovered latest offset was
not "normalized" and was running into the same issue.

I've moved the normalization at the read level of `ShapeStatus`, as I
think that's the cleanest boundary to place it at.

The crash recovery test + a small unit test cover the scenario.