feat: better buffering on move-ins (#3525)
New buffering means that instead of buffering all transactions until
move-in snapshot becomes known and then second time until results become
available, essentially rebasing the stream on the move-in snapshot, we
now do the opposite: rebase the move-in snapshot on whatever position
the stream happens to be in. We achieve this by keeping track of any
keys touched within a given snapshot window, and skipping those keys in
the move-in snapshot because those changes have already been made
visible from the stream.

What this achieves is we're buffering much less data (only keys for
inserts/updates instead of full transaction bodies) and not blocking any
consumers (which allows us to have 3+ layers of subqueries without
consistency issues).

Aside from that, this PR also fixes:
- Issue with storage where reads near the start of the stream could
yield a duplicate value (not reported before, but happened in tests for
move-outs)
- Issue with shadowing where some operations that happened before a
move-in was resolved were made visible when they shouldn't have
- Issue where a move-in followed by a move-out of the same value was
still appended to the log