fix: Handle out of bounds requests with a timeout (#3563)
Fixes https://github.com/electric-sql/electric/issues/3551 on top of
https://github.com/electric-sql/electric/pull/3562

I had pinpointed that there are cases in which a client might see a
response with offset B from one Electric, but that one has not been
flushed yet, and instead has only flushed the previous offset A.

At that point if there is a crash or handover, it is likely that the new
Electric recovers the earlier offset A, and would eventually pick up the
transaction with offset B from the replication slot. However when the
client tries to transparently continue its shape with the new Electric
with offset B, they occasionally run into an out of bounds error until
the new Electric catches up on the replication slot.

My initial intended solution was to simply treat offsets higher than the
latest seen from electric as offsets we hold in a live request, but
after some discussions with @balegas we've concluded to handle them as
errors after a timeout instead, allowing some time for Electric to catch
up and potentially rescue these requests with offsets higher than what
we've seen, but otherwise returning an error.

The idea would be that the error is also retryable, in which case we
could skip the timeout altogether, but that would require changing the
HTTP semantics of the API, and rather than return 400 non-retryable
errors, return some sort of 5xx error (and basically always assume the
client is right and we just don't have the offset they requested yet).


## Approach

I've first enhanced our rolling deploy integration test such that it
consistently reproduced this issue, by shifting a client from one
electric to the other mid-transaction ingestion.

Then I updated our API to no longer error when offsets are higher than
the last seen, and instead continue as they would normally, but mark
them as **_not up to date_**, and when it's time to serve the log,
_always_ try to handle them as live requests, but also with a message
sent to the request process itself after a timeout (half the long poll
timeout) to error if the offset has not been caught up to yet.

Another significant change I've made is to the `get_log_stream` storage
API, it now accepts a reverse range (the `up_to` being lower than the
`since`) and returns an empty enum, rather than error. If we think this
is not something we're ok with, I will have to rewire a few things -
honestly the only place this is a pain to fix is for SSE events.

One final thing to consider is also that since we know request processes
are long lived, we might need to be manually flushing/cleaning up the
message queues at the start of each request.

---------

Co-authored-by: Oleksii Sholik <oleksii@sholik.dev>