chore: Upgrade node to v24 (#3580)
We need NPM CLI >= 11.5.1 for OIDC to work [as per the Trusted Publisher
docs](https://docs.npmjs.com/trusted-publishers) which comes bundled
with node v24

Upgrading node to v24 caused some issues with JSDOM and the fetch
polyfills - which can be fixed by upgrading vitest to the latest
version.

I've refrained from upgrading any other deps to not complicate this PR
any further. This is the minimum viable thing to upgrade it.