feat: Reduce off heap allocations in `ShapeLogCollector` (#3585)
After an investigation with @alco on memory allocations, we've observed
that there's lots of strings being allocated off-heap by the SLC.

This is generally not an issue for most processes, but for a
high-frequency, performance sensitive process like SLC we want to
minimize unnecessary work and allocations and be extra careful about
what code executes in the very tight loop.

I've painstakingly gone through every allocation happening by the
process in the various steps of consuming a transaction, and have
identified 3 places where we were allocating things when we didn't need
to.

The main culprit is string interpolation, which allocates an
intermediate representation to perform the interpolation.

1. In our `OpenTelemetry.start_interval`, we pass a string, and later on
we do a string interpolation and atom creation to store the attribute.
We do this _for every_ interval type, _on every_ transaction fragment
handling.
2. The flush tracker notifies the replication client, but we were
calling `ReplicationClient.name` on _every_ notification call, which
does a string interpolation + atom cast.
3. Updating the `LsnTracker` ETS table was using the stack id, which
does a string interpolation for getting the table name.

These would generally be ok but at the rate at which they occur and for
how long lived the process is, these are potentially impacting
performance.

The solutions were easy:
1. Simply specify interval names as atoms, so we don't dynamically
generate any (which is safer) and we don't incur the cost of string
literals and concatenations - all compile time stuff.
2. Very simple, just generate the name at process initialization and
pass the atom to the closure rather than generate it in the closure
3. A bit annoying, but make the API accept a `stack_ref` which can be
either the stack ID or the table name, and add a `table_name` utility in
order to optimize the calls for this.